all edits to arduino.cpp and arduino.h within the arduino_sk_testing folder. 1) increased nump paterns from 12 to 100 2) added error_in_communication_2 and _3 to help debug an issue with loadsequence 3) a) added a static var LED1_intensity_, b) added a new property LED_1_dim with 0-100 range (should be remaned to intensity) and created a new property action called ondim (should rename to onChangeIntensity) c) implemented onDim method. if eAct is beforeGet then simply 'return' the static var LED1_intensity. if eAct is afterSet, then get the value of intensity that the property was changed to, then check if the port is clear and ready for com, then send char commands [23,1,X] whwre 23 is the flag to change intensity on the arduino firmware, 1 is supposed to be change intensity of LED1 and where x is the intensity value.
This code used chunks nspired from arduino communication in other places in the file and also from asiFW1000 change speed setting property. Propably many ways to implement this. currently the changing od LED1 is hardcoded. not sure if 1 function 'onDim" can hanndle responding to 3 different properties (one for each LED) by also recieving the led number from the property name and sending that along as the 2nd char in the command to arduino. or if there needs to be 3 different action functions oneDim1, ondim2 odim3... each linked to the respective property.